Moved DynamicConfig into Installer
