+++ protected/humhub/modules/installer/controllers/SetupController.php
@@ -11,9 +11,9 @@ namespace humhub\modules\installer\controllers;
 use Exception;
 use humhub\components\access\ControllerAccess;
 use humhub\components\Controller;
-use humhub\libs\DynamicConfig;
 use humhub\modules\admin\widgets\PrerequisitesList;
 use humhub\modules\installer\forms\DatabaseForm;
+use humhub\modules\installer\libs\DynamicConfig;
 use humhub\modules\installer\Module;
 use humhub\services\MigrationService;
 use Yii;
